C17 : Weazel News
Adjusted pay to weazel news to allow for some spending money to pay for tips. There is someone running weazel news and providing RP for it. 
Added Role for weazel news(freelancer). I'm not sure how the job center works, but the weazel news owner wanted to still allow people to be able to do the job from the job center as a freelancer and allow him to purchase what they report on.